chore: disable broken tests
