+++ src/components/ItaliaTheme/View/__disabled_tests__/BandoView.test__disabled.jsx
