+++ src/components/ItaliaTheme/View/__disabled_tests__/EventoView.test__disabled.jsx
