+++ src/components/ItaliaTheme/View/__disabled_tests__/NewsItemView.test__disabled.jsx
