+++ src/components/ItaliaTheme/View/__disabled_tests__/PuntoDiContattoView.test_disabled.jsx
