+++ src/components/ItaliaTheme/View/__disabled_tests__/VenueView.test_disabled.jsx
